refactor(scheduler): share REST DTOs and extract graph-to-DTO builders
Move the scheduler REST API's JobResponse/TaskSummary/TaskStatus/Percentiles/
QueryStageSummary/QueryStagesResponse types onto the ballista-history crate's
shared DTOs, and extract the graph-to-DTO construction logic that previously
lived inline in the handlers into pub(crate) builder functions in a new
api::dto_build module. Handlers become thin wrappers over these builders,
which a later event-log writer will also call. Behavior is unchanged; the
existing handler unit tests move to dto_build.rs alongside the moved helper
functions they exercise, plus one new plan-string assertion test.
