+++ ballista/scheduler/Cargo.toml
@@ -51,6 +51,7 @@ async-trait = { workspace = true }
 axum = "0.8.9"
 tower-http = { version = "0.7", features = ["cors"] }
 ballista-core = { path = "../core", version = "53.0.0" }
+ballista-history = { path = "../history", version = "53.0.0" }
 clap = { workspace = true, optional = true }
 dashmap = { workspace = true }
 datafusion = { workspace = true }
@@ -81,6 +82,7 @@ uuid = { workspace = true }
 
 [dev-dependencies]
 rstest = { workspace = true }
+serde_json = "1"
 
 [build-dependencies]
 tonic-prost-build = { workspace = true, optional = true }
